vtls: move trust management routines from sectransp.c to apple.c
